Move save_model and load_model to utils to avoid cyclic imports
